rag-openai/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-cloud-aiplatform>=1.91.0",
    "google-cloud-discoveryengine>=0.13.8",
    "langchain>=0.3.24",
    "langchain-community>=0.3.23",
    "langchain-google-community>=2.0.7",
    "langchain-google-genai>=2.1.4",
    "load-dotenv>=0.1.0",
    "openai>=1.77.0",
    "pip>=25.1.1",
    "uvicorn>=0.34.2",
]
